Add WiringPi wrapper module
Add a wrapper module for the wiringpi library http://wiringpi.com.

WiringPi is a GPIO access library written in C for the BCM2835 used in the Raspberry Pi. It’s released under the GNU LGPLv3 license and is usable from C and C++ and many other languages with suitable wrappers. It’s designed to be familiar to people who have used the Arduino “wiring” system.